fix circular dep in gatsby-link
