+++ packages/gatsby-link/src/index.js
@@ -4,38 +4,9 @@ import { Link, Location } from "@gatsbyjs/reach-router"
 import { parsePath } from "./parse-path"
 import { isLocalLink } from "./is-local-link"
 import { rewriteLinkPath } from "./rewrite-link-path"
+import { withPrefix, getGlobalPathPrefix } from "./prefix-helpers"
 
-export { parsePath }
-
-export function withPrefix(path, prefix = getGlobalBasePrefix()) {
-  if (!isLocalLink(path)) {
-    return path
-  }
-
-  if (path.startsWith(`./`) || path.startsWith(`../`)) {
-    return path
-  }
-  const base = prefix ?? getGlobalPathPrefix() ?? `/`
-
-  return `${base?.endsWith(`/`) ? base.slice(0, -1) : base}${
-    path.startsWith(`/`) ? path : `/${path}`
-  }`
-}
-
-// These global values are wrapped in typeof clauses to ensure the values exist.
-// This is especially problematic in unit testing of this component.
-const getGlobalPathPrefix = () =>
-  process.env.NODE_ENV !== `production`
-    ? typeof __PATH_PREFIX__ !== `undefined`
-      ? __PATH_PREFIX__
-      : undefined
-    : __PATH_PREFIX__
-const getGlobalBasePrefix = () =>
-  process.env.NODE_ENV !== `production`
-    ? typeof __BASE_PATH__ !== `undefined`
-      ? __BASE_PATH__
-      : undefined
-    : __BASE_PATH__
+export { parsePath, withPrefix }
 
 export function withAssetPrefix(path) {
   return withPrefix(path, getGlobalPathPrefix())
